Finalized and moved function/class declarations to separate files
